Module re_chunk_store::dataframe

All the APIs used specifically for re_dataframe.

Structs

ComponentColumnDescriptor

Describes a data/component column, such as Position3D.

ComponentColumnSelector

Select a component based on its EntityPath and ComponentName.

QueryExpression

Describes a complete query for Rerun’s dataframe API.

TimeColumnDescriptor

Describes a time column, such as log_time.

TimeColumnSelector

Select a time column.

Enums

ColumnDescriptor

ColumnSelector

Describes a column selection to return as part of a query.

SparseFillStrategy

Specifies how null values should be filled in the returned dataframe.

Type Aliases

Index

IndexRange

IndexValue

ViewContentsSelector

The view contents specify which subset of the database (i.e., which columns) the query runs on, expressed as a set of [EntityPath]s and their associated [ComponentName]s.