Crate re_chunk_store

The Rerun chunk store, implemented on top of Apache Arrow using the [arrow] crate.

This crate is an in-memory time series database for Rerun log data. It is indexed by Entity path, component, timeline, and time. It supports out-of-order insertions, and fast O(log(N)) queries.

See ChunkStore for an overview of the core data structures.

See ChunkStore::latest_at_relevant_chunks and ChunkStore::range_relevant_chunks for the documentation of the public read APIs.

See ChunkStore::insert_chunk for the documentation of the public write APIs.

Feature flags

deadlock_detection — Enables parking_lot’s deadlock detection background thread.

Re-exports

pub use re_chunk::Chunk;

pub use re_chunk::ChunkId;

pub use re_chunk::ChunkShared;

pub use re_chunk::LatestAtQuery;

pub use re_chunk::RangeQuery;

pub use re_chunk::RangeQueryOptions;

pub use re_chunk::RowId;

pub use re_chunk::UnitChunkShared;

pub use re_log_types::ResolvedTimeRange;

pub use re_log_types::TimeInt;

pub use re_log_types::TimeType;

pub use re_log_types::Timeline;

Modules

dataframe 🔒

All the APIs used specifically for re_dataframe.

drop_time_range 🔒

events 🔒

external

gc 🔒

query 🔒

stats 🔒

store 🔒

subscribers 🔒

writes 🔒

Structs

ChunkCompactionReport

Reports which [Chunk]s were merged into a new [Chunk] during a compaction.

ChunkStore

A complete chunk store: covers all timelines, all entities, everything.

ChunkStoreChunkStats

Stats about a collection of chunks.

ChunkStoreConfig

ChunkStoreDiff

Describes an atomic change in the Rerun ChunkStore: a chunk has been added or deleted.

ChunkStoreEvent

The atomic unit of change in the Rerun ChunkStore.

ChunkStoreGeneration

Incremented on each edit.

ChunkStoreHandle

A ref-counted, inner-mutable handle to a ChunkStore.

ChunkStoreStats

ChunkStoreSubscriberHandle

ColumnMetadata

ComponentColumnDescriptor

Describes a data/component column, such as Position3D, in a dataframe.

GarbageCollectionOptions

IndexColumnDescriptor

Describes a time column, such as log_time.

QueryExpression

Describes a complete query for Rerun’s dataframe API.

ViewContentsSelector

The view contents specify which subset of the database (i.e., which columns) the query runs on, expressed as a set of [EntityPath]s and their associated [ComponentName]s.

Enums

ChunkStoreDiffKind

Is it an addition or a deletion?

ChunkStoreError

ColumnDescriptor

Describes any kind of column.

GarbageCollectionTarget

SparseFillStrategy

Specifies how null values should be filled in the returned dataframe.

Traits

ChunkStoreSubscriber

A ChunkStoreSubscriber subscribes to atomic changes from all ChunkStores through ChunkStoreEvents.

PerStoreChunkSubscriber

A ChunkStoreSubscriber that is instantiated for each unique [StoreId].

Type Aliases

ChunkStoreResult

Index

IndexRange

IndexValue